old-tests/submission/W3C/domhtml/HTMLIFrameElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-iframe-align" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe align="top" src="/support/html/sample.html" height='80' width="550">IFRAME1</iframe>


<script>
var domtest = async_test("The align IDL attribute of the iframe element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var valign;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      valign = testNode.align;

      assert_equals(valign,"top");
       
    domtest.done();
}));

</script>
<p>The align IDL attribute of the iframe element must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
